fix(Grid): remove externalClasses
