+++ packages/products/tdesign-miniprogram/src/grid/README.en-US.md
@@ -21,25 +21,6 @@ className | Description
 t-class | \-
 
 
-### Col Props
-
-name | type | default | description | required
--- | -- | -- | -- | --
-style | Object | - | CSS(Cascading Style Sheets) | N
-custom-style | Object | - | CSS(Cascading Style Sheets)，used to set style on virtual component | N
-offset | String / Number | - | \- | N
-span | String / Number | - | \- | N
-
-
-### Row Props
-
-name | type | default | description | required
--- | -- | -- | -- | --
-style | Object | - | CSS(Cascading Style Sheets) | N
-custom-style | Object | - | CSS(Cascading Style Sheets)，used to set style on virtual component | N
-gutter | String / Number | - | \- | N
-
-
 ### GridItem Props
 
 name | type | default | description | required
