+++ packages/products/tdesign-miniprogram/src/grid/README.md
@@ -21,25 +21,6 @@ theme | String | default | 宫格的风格。可选项：default/card | N
 t-class | 根节点样式类
 
 
-### Col Props
-
-名称 | 类型 | 默认值 | 描述 | 必传
--- | -- | -- | -- | --
-style | Object | - | 样式 | N
-custom-style | Object | - | 样式，一般用于开启虚拟化组件节点场景 | N
-offset | String / Number | - | 列的偏移量（默认单位px） | N
-span | String / Number | - | 列的宽度（默认单位px） | N
-
-
-### Row Props
-
-名称 | 类型 | 默认值 | 描述 | 必传
--- | -- | -- | -- | --
-style | Object | - | 样式 | N
-custom-style | Object | - | 样式，一般用于开启虚拟化组件节点场景 | N
-gutter | String / Number | - | 列之间的间距（默认单位px） | N
-
-
 ### GridItem Props
 
 名称 | 类型 | 默认值 | 描述 | 必传
